feat(discover): move widgets in discover page into separated file
